requirements.txt
asgiref==3.4.1
cffi==1.15.0
cheroot==8.6.0
CherryPy==18.6.1
cryptography==36.0.1
Django==3.2.11
django-ckeditor==6.2.0
django-jasmine==0.5.1
django-jazzmin==2.4.8
django-js-asset==1.2.2
djangorestframework==3.13.1
glob2==0.7
importlib-resources==5.4.0
jaraco.classes==3.2.1
jaraco.collections==3.4.0
jaraco.functools==3.4.0
jaraco.text==3.6.0
jasmine==3.99.0
jasmine-core==3.99.0
Jinja2==2.11.3
MarkupSafe==2.0.1
more-itertools==8.12.0
ordereddict==1.1
Pillow==8.4.0
portend==3.0.0
pycparser==2.21
PyMySQL==1.0.2
pytz==2021.3
PyYAML==6.0
selenium==3.141.0
six==1.16.0
sqlparse==0.4.2
tempora==4.1.2
typing_extensions==4.0.1
urllib3==1.26.8
zc.lockfile==2.0
zipp==3.6.0
